+++ draft-packages/tile/docs/Tile.stories.tsx
@@ -10,11 +10,18 @@ import {
 import { Coaching } from "@kaizen/draft-illustration"
 import { Tag } from "@kaizen/draft-tag"
 import { Paragraph } from "@kaizen/component-library"
-
+import { withDesign } from "storybook-addon-designs"
 import bookmarkIcon from "@kaizen/component-library/icons/bookmark-off.icon.svg"
+import { figmaEmbed } from "../../../storybook/helpers"
 
 export default {
   title: "Tile (React)",
+  parameters: {
+    ...figmaEmbed(
+      "https://www.figma.com/file/GMxm8rvDCbj0Xw3TQWBZ8b/UI-Kit-Zen?node-id=14489%3A69120"
+    ),
+  },
+  decorators: [withDesign],
 }
 
 const primaryAction: TileAction = {