+++ draft-packages/well/docs/Well.stories.tsx
@@ -2,6 +2,8 @@ import { Box, Heading, Paragraph } from "@kaizen/component-library"
 import { TextField } from "@kaizen/draft-form"
 import { Well } from "@kaizen/draft-well"
 import * as React from "react"
+import { withDesign } from "storybook-addon-designs"
+import { figmaEmbed } from "../../../storybook/helpers"
 
 const ExampleContent = () => (
   <div style={{ padding: "1em 2em", maxWidth: "400px" }}>
@@ -32,7 +34,11 @@ export default {
       import { Well } from "@kaizen/draft-well"
       `,
     },
+    ...figmaEmbed(
+      "https://www.figma.com/file/GMxm8rvDCbj0Xw3TQWBZ8b/UI-Kit-Zen?node-id=1929%3A14168"
+    ),
   },
+  decorators: [withDesign],
 }
 
 export const DefaultWithSolidBorderKaizenSiteDemo = () => (