+++ storybook/helpers/figmaEmbed.ts
@@ -1,0 +1,6 @@
+export const figmaEmbed = (url: string) => ({
+  design: {
+    type: "figma",
+    url,
+  },
+})