+++ draft-packages/badge/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.2.3...@kaizen/draft-badge@1.2.4) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-badge
+
+
+
+
+
 ## [1.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.2.2...@kaizen/draft-badge@1.2.3) (2021-01-11)
 
 **Note:** Version bump only for package @kaizen/draft-badge