+++ draft-packages/dropdown/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.62](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown@1.3.61...@kaizen/draft-dropdown@1.3.62) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown
+
+
+
+
+
+## [1.3.61](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown@1.3.60...@kaizen/draft-dropdown@1.3.61) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown
+
+
+
+
+
 ## [1.3.60](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown@1.3.59...@kaizen/draft-dropdown@1.3.60) (2021-01-18)
 
 **Note:** Version bump only for package @kaizen/draft-dropdown