+++ draft-packages/hierarchical-select/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.56](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-select@1.3.55...@kaizen/draft-hierarchical-select@1.3.56) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-select
+
+
+
+
+
+## [1.3.55](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-select@1.3.54...@kaizen/draft-hierarchical-select@1.3.55) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-select
+
+
+
+
+
 ## [1.3.54](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-select@1.3.53...@kaizen/draft-hierarchical-select@1.3.54) (2021-01-28)
 
 **Note:** Version bump only for package @kaizen/draft-hierarchical-select