+++ draft-packages/likert-scale-legacy/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.14](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-likert-scale-legacy@1.1.13...@kaizen/draft-likert-scale-legacy@1.1.14) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-likert-scale-legacy
+
+
+
+
+
+## [1.1.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-likert-scale-legacy@1.1.12...@kaizen/draft-likert-scale-legacy@1.1.13) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-likert-scale-legacy
+
+
+
+
+
 ## [1.1.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-likert-scale-legacy@1.1.11...@kaizen/draft-likert-scale-legacy@1.1.12) (2021-01-18)
 
 **Note:** Version bump only for package @kaizen/draft-likert-scale-legacy