+++ draft-packages/loading-placeholder/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-loading-placeholder@1.4.5...@kaizen/draft-loading-placeholder@1.4.6) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-loading-placeholder
+
+
+
+
+
+## [1.4.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-loading-placeholder@1.4.4...@kaizen/draft-loading-placeholder@1.4.5) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-loading-placeholder
+
+
+
+
+
 ## [1.4.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-loading-placeholder@1.4.3...@kaizen/draft-loading-placeholder@1.4.4) (2021-01-18)
 
 **Note:** Version bump only for package @kaizen/draft-loading-placeholder