+++ draft-packages/search-box/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.3...@kaizen/draft-search-box@1.1.4) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-search-box
+
+
+
+
+
+## [1.1.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.2...@kaizen/draft-search-box@1.1.3) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-search-box
+
+
+
+
+
 ## [1.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.1...@kaizen/draft-search-box@1.1.2) (2021-01-28)
 
 **Note:** Version bump only for package @kaizen/draft-search-box