+++ draft-packages/select/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.38](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.37...@kaizen/draft-select@1.11.38) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-select
+
+
+
+
+
+## [1.11.37](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.36...@kaizen/draft-select@1.11.37) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-select
+
+
+
+
+
 ## [1.11.36](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.35...@kaizen/draft-select@1.11.36) (2021-01-28)
 
 **Note:** Version bump only for package @kaizen/draft-select