+++ draft-packages/tag/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@1.7.1...@kaizen/draft-tag@1.7.2) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-tag
+
+
+
+
+
+## [1.7.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@1.7.0...@kaizen/draft-tag@1.7.1) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-tag
+
+
+
+
+
 # [1.7.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@1.6.4...@kaizen/draft-tag@1.7.0) (2021-01-19)