+++ draft-packages/user-interactions/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.91](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-user-interactions@1.2.90...@kaizen/draft-user-interactions@1.2.91) (2021-02-05)
+
+**Note:** Version bump only for package @kaizen/draft-user-interactions
+
+
+
+
+
+## [1.2.90](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-user-interactions@1.2.89...@kaizen/draft-user-interactions@1.2.90) (2021-02-02)
+
+**Note:** Version bump only for package @kaizen/draft-user-interactions
+
+
+
+
+
 ## [1.2.89](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-user-interactions@1.2.88...@kaizen/draft-user-interactions@1.2.89) (2021-01-18)
 
 **Note:** Version bump only for package @kaizen/draft-user-interactions